requirements.txt
openai
requests
geographiclib
networkx
prompt_toolkit
rich
python-dotenv
pydantic